Add GraphQL data type and query schema
Issue: Add schema for query and data type in server to be used to communicate with client.
Solution: Add habit, identity schema and queries in schema.js and connect this to GraphQL middleware for './data' url
Risk: Data communication has only been tested with Graphiql Client. Further testing is needed for communication with actual client.
Reviewed by: Jason Ko