Connect MongoDB Atlas to GraphQL server
Issue: There was no Database used for the server to communitate
Solution: Use MongoDB Atlas as the database and change all the resolvers to communicate with Atlas
Risk: When adding new habit with invalid identity id, it currently does not return any error. It will simply not add the habit to the database but the client will not know. This can be fixed later.
Reviewed by: Jason Ko